split ChetSession to separate file
